Extract cli_readline.go
